heated build plate json is a float

+++ src/model/bot_model.h
@@ -165,7 +165,7 @@ class BotModel : public BaseModel {
     MODEL_PROP(int, buildplaneCurrentTemp, -999)
     MODEL_PROP(int, buildplaneTargetTemp, -999)
     MODEL_PROP(int, chamberErrorCode, 0)
-    MODEL_PROP(int, hbpCurrentTemp, -999)
+    MODEL_PROP(float, hbpCurrentTemp, -999.0f)
     MODEL_PROP(int, hbpTargetTemp, -999)
     MODEL_PROP(int, filamentBayATemp, -999)
     MODEL_PROP(int, filamentBayBTemp, -999)
